sri_lanka_trip_planner/pyproject.toml
[project]
name = "sri_lanka_trip_planner"
version = "0.1.0"
description = "sri_lanka_trip_planner using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.14.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3"
]

[project.scripts]
sri_lanka_trip_planner = "sri_lanka_trip_planner.main:run"
run_crew = "sri_lanka_trip_planner.main:run"
train = "sri_lanka_trip_planner.main:train"
replay = "sri_lanka_trip_planner.main:replay"
test = "sri_lanka_trip_planner.main:test"
run_with_trigger = "sri_lanka_trip_planner.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
